changed some constants to doubles so the pow() template function doesn't regard these calls as ambiguous...

+++ CLAM/src/Processing/Synthesis/SynthSineSpectrum.cxx
@@ -677,7 +677,7 @@ bool SynthSineSpectrum::Do(const SpectralPeakArray& peakArray,Spectrum& residual
 
 void SynthSineSpectrum::FillSynthSineSpectrum	(	const SpectralPeakArray& peakArray, double gain)
 {
-	TSize 		mainLobeBins = TData(8)*pow(2,mConfig.GetZeroPadding()) ; // Number of bins in the mainlope of a transformed BHarris92 window 
+	TSize 		mainLobeBins = TData(8)*pow(2.0,(double)mConfig.GetZeroPadding()) ; // Number of bins in the mainlope of a transformed BHarris92 window 
                            
 	TIndex 		b,i,k,l,Index,numPeaks = peakArray.GetnPeaks();
   					
@@ -697,14 +697,14 @@ void SynthSineSpectrum::FillSynthSineSpectrum	(	const SpectralPeakArray& peakArr
 	TData binPosFactor=2*(spectrumSize-1)/samplingRate;
 	TData firstBinFactor= 3* pow((TData)2,(TData)zeroPadding);
 
-	int incr=int(pow(2,9-mConfig.GetZeroPadding()));
+	int incr=int(pow(2.0,9.0-mConfig.GetZeroPadding()));
 	int lastBin=4096-incr;
 	/* loop thru all the peaks ...*/
 	for(i=0;i<numPeaks;i++)
 	{
 		if(peakArray.GetScale()==EScale::eLog)//if it is in dB
 		{
-			currMag = pow(10,(peakMagBuffer[i]/20));
+			currMag = pow(10.0,(peakMagBuffer[i]/20.0));
 		}
 		else	currMag = peakMagBuffer[i];
 		currFreq = peakFreqBuffer[i];
